+++ src/Aks/Aks.Test/ScenarioTests/KubernetesTests.cs
@@ -51,5 +51,12 @@ namespace Commands.Aks.Test.ScenarioTests
         {
             TestController.NewInstance.RunPowerShellTest(_logger, "Test-NewAzAksWithAcr");
         }
+        
+        [Fact(Skip = "Updating service principal profile is not allowed on MSI cluster.")]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestResetAzureKubernetesServicePrincipal()
+        {
+            TestController.NewInstance.RunPowerShellTest(_logger, "Test-ResetAzureKubernetesServicePrincipal");
+        }
     }
 }