+++ src/Aks/Aks/ChangeLog.md
@@ -20,6 +20,7 @@
 ## Upcoming Release
 * Added support `AcrNameToAttach` in `Set-AzAksCluster`. [#14692]
 * Added support `AcrNameToDetach` in `Set-AzAksCluster`. [#14693] 
+* Added `Set-AzAksClusterCredential` to reset the ServicePrincipal of an existing AKS cluster.
 
 ## Version 2.0.2
 * Refined error messages of cmdlet failure.